Split marm-console's Memory.tsx into per-tab component files
Pure refactor, no behavior changes. Extracted via exact index-based
line-slicing (all 4 moved bodies verified byte-identical to the
original before any import cleanup) -- same methodology as the
memory_ops.py split.

- src/pages/Memory.tsx keeps only the MemoryPage shell: 1075 -> 53 lines.
- src/components/memory/shared.tsx (new, 52 lines): ActionNotice type,
  mutationErrorMessage, deleteNotice, ActionNoticePanel -- used by all
  5 tabs.
- src/components/memory/MemoriesTab.tsx (new, 403 lines): MemoryRow +
  MemoriesTab.
- src/components/memory/SessionsAndLogsTabs.tsx (new, 297 lines):
  SessionSummaryDialog + SessionsTab + LogsTab.
- src/components/memory/NotebookAndCompactionTabs.tsx (new, 289
  lines): HISTORY_STATUS_LABEL + CompactionHistoryRow + NotebookTab +
  CompactionTab.

Split granularity (MemoriesTab alone, smaller tabs grouped in 2 files)
confirmed with user directly before implementation. New
components/memory/ directory mirrors the existing components/layout/
and components/ui/ convention, chosen over pages/memory/ specifically
to avoid a case-only collision with pages/Memory.tsx on case-insensitive
filesystems.

Each extracted component referenced across file boundaries gained
`export` (mechanical requirement -- TS won't let you import an
unexported binding); components that stayed file-internal (MemoryRow,
SessionSummaryDialog, CompactionHistoryRow) were left unexported,
same visibility as before.

Validated: tsc --noEmit clean, npm run build succeeds, and a real
Playwright smoke pass against the running dev server confirmed all 5
tabs render and are clickable with zero JS console errors.
